Renamed app to tutorialApp since previous name is no longer accurate
